discovery/src/requirements.txt
cloudshell-shell-core~=5.1
attrs
